TASK-07,08,09: Added command line interface, facade service to manage products and invoice date and time.

+++ src/main/java/service/impl/ComputerServiceImpl.java
@@ -1,39 +1,88 @@
 package service.impl;
 
 import dto.product.request.ComputerRequestDto;
+import dto.product.request.ProductRequestDto;
 import dto.product.response.ComputerResponseDto;
+import entity.client.Client;
+import entity.client.Role;
 import entity.product.type.Computer;
+import exception.ClientNotFoundException;
 import exception.ComputerNotFoundException;
+import exception.NoPermissionsException;
 import lombok.RequiredArgsConstructor;
 import mapper.ComputerMapper;
-import repository.ComputerRepository;
+import repository.ClientRepository;
+import repository.productrepositories.ComputerRepository;
 import service.ComputerService;
-import utils.ProductIdGenerator;
+import repository.productrepositories.idconfig.ProductIdGenerator;
+import validator.ProductValidator;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RequiredArgsConstructor
 public class ComputerServiceImpl implements ComputerService {
     private final ComputerRepository computerRepository;
-    private final ProductIdGenerator productIdGenerator;
+    private final ClientRepository clientRepository;
 
     @Override
-    public ComputerResponseDto create(ComputerRequestDto computerRequestDto) {
+    public ComputerResponseDto create(ProductRequestDto dto, Long clientId) {
+        ProductValidator.validate(dto);
+
+        ComputerRequestDto computerRequestDto = (ComputerRequestDto) dto;
+
+        Client client = clientRepository.getClientById(clientId).orElseThrow(ClientNotFoundException::new);
+
+        if (client.getRole() != Role.ADMIN) {
+            throw new NoPermissionsException();
+        }
+
         Computer computer = computerRepository.save(ComputerMapper.mapDtoToComputer(computerRequestDto,
-                productIdGenerator.getNextProductId()));
+                ProductIdGenerator.getNextProductId()));
         return ComputerMapper.mapComputerToDto(computer);
     }
 
     @Override
-    public ComputerResponseDto remove(Long id) {
+    public ComputerResponseDto remove(Long id, Long clientId) {
+        Client client = clientRepository.getClientById(clientId).orElseThrow(ClientNotFoundException::new);
+
+        if (client.getRole() != Role.ADMIN) {
+            throw new NoPermissionsException();
+        }
+
         Computer computer = computerRepository.delete(id).orElseThrow(ComputerNotFoundException::new);
         return ComputerMapper.mapComputerToDto(computer);
     }
 
     @Override
-    public ComputerResponseDto update(Long id, ComputerRequestDto computerRequestDto) {
-        Computer computer = computerRepository.update(id, ComputerMapper.mapDtoToComputer(computerRequestDto, id))
+    public ComputerResponseDto updatePrice(Long id, Long clientId, BigDecimal price) {
+        ProductValidator.validatePrice(price);
+
+        Client client = clientRepository.getClientById(clientId).orElseThrow(ClientNotFoundException::new);
+
+        if (client.getRole() != Role.ADMIN) {
+            throw new NoPermissionsException();
+        }
+
+        Computer computer = computerRepository.updateComputerPrice(id, price)
                 .orElseThrow(ComputerNotFoundException::new);
+
+        return ComputerMapper.mapComputerToDto(computer);
+    }
+
+    @Override
+    public ComputerResponseDto updateQuantity(Long id, Long clientId, int quantity) {
+        ProductValidator.validateQuantity(quantity);
+
+        Client client = clientRepository.getClientById(clientId).orElseThrow(ClientNotFoundException::new);
+
+        if (client.getRole() != Role.ADMIN) {
+            throw new NoPermissionsException();
+        }
+
+        Computer computer = computerRepository.updateComputerQuantity(id, quantity)
+                .orElseThrow(ComputerNotFoundException::new);
+
         return ComputerMapper.mapComputerToDto(computer);
     }
 
@@ -48,5 +97,15 @@ public class ComputerServiceImpl implements ComputerService {
         return computerRepository.getAllComputers().stream()
                 .map(ComputerMapper::mapComputerToDto)
                 .toList();
+    }
+
+    @Override
+    public boolean exist(Long id) {
+        return computerRepository.getComputerById(id).isPresent();
+    }
+
+    @Override
+    public boolean isInstance(ProductRequestDto dto) {
+        return dto instanceof ComputerRequestDto;
     }
 }
